requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
amqp==5.3.1
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
black==22.12.0
bleach==6.1.0
boto3==1.21.4
botocore==1.24.4
celery==5.2.3
certifi==2021.10.8
cffi==1.17.1
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==2.3.0
Django==4.2.6
django-auto-logout==0.5.1
django-celery-beat==2.2.1
django-celery-results==2.5.1
django-ckeditor==6.2.0
django-cors-headers==4.7.0
django-environ==0.11.2
django-filter==25.1
django-graphql-jwt==0.4.0
django-js-asset==2.0.0
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.9
executing==2.0.1
fastjsonschema==2.19.1
flex==6.14.1
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.5.0
graphene==3.4.3
graphene-django==3.2.2
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jdatetime==3.8.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kavenegar==1.1.2
kiwisolver==1.4.5
kombu==5.2.3
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.1.3
notebook_shim==0.2.4
npm==0.1.1
numpy==1.26.4
optional-django==0.1.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdfminer==20191125
pillow==11.1.0
platformdirs==4.2.1
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.28
propcache==0.2.0
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.17.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.0.7
PyPDF2==3.0.1
python-crontab==2.6.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytse-client==0.18.0
pytz==2025.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
redis==4.1.4
referencing==0.35.0
regex==2024.9.11
requests==2.27.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rpds-py==0.18.0
s3transfer==0.5.1
scipy==1.14.0
Send2Trash==1.8.3
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.6.3
strict-rfc3339==0.7
sympy==1.13.3
tenacity==7.0.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tornado==6.4
tqdm==4.66.6
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.8
validate_email==1.3
vine==5.0.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wrapt==1.14.0
yarl==1.17.1
graphene-file-upload